# core, eth/downloader: commit block data using batches