In this PR, fast sync state and receipt data is committed in batches of ~100kB. This size was determined using experiments with https://github.com/fjl/goleveldb-bench. Another related optimization in this PR removes random reads in HasBlock and HasHeader.

The PR also restores batch writes of state data for full imports and writes the block, its receipts and most other data in the same batch.