@fjl 2017-09-07
It's really hard to measure the impact of this change. Package core benchmarks show a slight improvement for disk databases.

```
name                                old time/op    new time/op    delta
InsertChain_empty_memdb-8             95.5µs ± 1%    96.8µs ± 1%     ~     (p=0.095 n=5+5)
InsertChain_empty_diskdb-8             265µs ± 5%     209µs ± 1%  -21.02%  (p=0.008 n=5+5)
InsertChain_valueTx_memdb-8            185µs ± 2%     180µs ± 1%   -2.69%  (p=0.008 n=5+5)
InsertChain_valueTx_diskdb-8           379µs ± 1%     296µs ± 2%  -21.90%  (p=0.008 n=5+5)
InsertChain_valueTx_100kB_memdb-8      980µs ± 1%     998µs ± 4%     ~     (p=0.111 n=4+5)
InsertChain_valueTx_100kB_diskdb-8    1.42ms ± 1%    1.34ms ± 4%   -5.84%  (p=0.008 n=5+5)
InsertChain_uncles_memdb-8             195µs ± 2%     196µs ± 1%     ~     (p=0.690 n=5+5)
InsertChain_uncles_diskdb-8            360µs ± 1%     306µs ± 1%  -14.82%  (p=0.008 n=5+5)
InsertChain_ring200_memdb-8           84.9ms ± 2%    84.6ms ± 1%     ~     (p=1.000 n=5+5)
InsertChain_ring200_diskdb-8          83.2ms ± 1%    78.6ms ± 1%   -5.52%  (p=0.008 n=5+5)
InsertChain_ring1000_memdb-8          89.9ms ± 3%    88.7ms ± 1%     ~     (p=0.310 n=5+5)
InsertChain_ring1000_diskdb-8         89.0ms ± 1%    82.9ms ± 1%   -6.81%  (p=0.008 n=5+5)
```